requirements.txt
Flask==3.0.0
SQLAlchemy==1.4.50
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Pillow==11.0.0
WTForms==3.1.1
opencv-python>=4.5.0
numpy>=1.19.0
sympy>=1.12
